Move LoadingScreen.jsx to file-only component
