--- src/browser/components/LoadingScreen/index.js
@@ -1,4 +1,0 @@
-// Copyright (c) 2016-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-
-export {default} from './LoadingScreen.jsx';
